vaults move: ../data_sources/[dsId]/tables/[tId]/rows/index
